Refactored - moved file reading/moving/logging from email_monitoring&autoImportFromFTP into a new module, ServerFiles, for use in the compare script
