+++ .gitignore
@@ -31,3 +31,5 @@ dist*
 client_secrets.json
 token.json
 sample-form-reader-395ab8f59999.json
+
+Temp Additions/
